requirements.txt
appdirs==1.4.4
autograd==1.4
autoray==0.3.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
cycler==0.11.0
dill==0.3.5.1
fonttools==4.33.3
future==0.18.2
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
mpmath==1.2.1
networkx==2.8.3
ninja==1.10.2.3
ntlm-auth==1.5.0
numpy==1.22.4
packaging==21.3
pandas==1.4.3
pbr==5.9.0
PennyLane==0.23.1
PennyLane-Lightning==0.23.0
Pillow==9.1.1
ply==3.11
psutil==5.9.1
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
qiskit==0.37.1
qiskit-aer==0.10.4
qiskit-ibmq-provider==0.19.2
qiskit-terra==0.21.1
requests==2.28.1
requests-ntlm==1.1.0
retworkx==0.11.0
scikit-learn==1.1.1
scipy==1.8.1
semantic-version==2.6.0
six==1.16.0
sklearn==0.0
stevedore==4.0.0
symengine==0.9.2
sympy==1.10.1
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
tweedledum==1.1.1
urllib3==1.26.11
websocket-client==1.3.3
websockets==10.3
